lce: make sure MM transactions are shutdown last
Insert MM transactions at the end of the list to make sure they get
destroyed last on shutdown. This makes sure that other protocols
which might invoke and wait for the completion of MM transactions
have their transactions terminated first and don't mistake a link
shutdown or a protocol specific error for a MM error.

Signed-off-by: Patrick McHardy <kaber@trash.net>