netlink: restore broadcast reception handler after sending unicast requests
Instead of setting the handler to NULL, restore the broadcast message
reception handler. Otherwise messages will be lost after sending a
unicast request. This currently only works because the requests are
either sent before the broadcast message handler is installed by the
initialization function or (in FP mode) there are no broadcast messages.

Signed-off-by: Patrick McHardy <kaber@trash.net>